pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["one-shot-monkey"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.2.0"
click = "^8.1.7"
pyyaml = "^6.0.1"
typer = "^0.12.3"
ipython = "^8.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
